# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PHP Teste — API de Propostas",
        "description": "API REST para gestão de clientes, propostas financeiras e pedidos. Implementa máquina de estados, optimistic locking, idempotência e auditoria automática.",
        "contact": {
            "name": "Higor Moreira",
            "email": "higor@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Lista pedidos do usuário autenticado",
                "description": "Lista os pedidos do usuário autenticado, com filtro opcional por status.",
                "operationId": "00d3856d2d3d93124e0cc021d0cb5b79",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de pedidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/propostas/{proposta}/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cria um pedido a partir de uma proposta aprovada",
                "description": "Cria um pedido a partir de uma proposta APPROVED.",
                "operationId": "6d037f518a3c5bc3c6b56ff85e4d4baf",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "observacoes": {
                                        "type": "string",
                                        "example": "Entrega urgente",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Proposta não está aprovada ou já possui pedido ativo"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Exibe um pedido",
                "description": "Exibe um pedido específico do usuário autenticado.",
                "operationId": "38713af501dbd66af5a66bd568a49bc2",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão para visualizar este pedido"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancela um pedido",
                "description": "Cancela um pedido.",
                "operationId": "aee443a742a80c8c038743d02cee6c66",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido cancelado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão para cancelar este pedido"
                    },
                    "422": {
                        "description": "Pedido não pode ser cancelado no status atual"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "proposta_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ],
                        "example": "pending"
                    },
                    "status_label": {
                        "type": "string",
                        "example": "Aguardando pagamento"
                    },
                    "valor_total": {
                        "type": "number",
                        "format": "float",
                        "example": 1200
                    },
                    "observacoes": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Informe o access_token obtido em POST /auth/login",
                "bearerFormat": "OAuth2 (Laravel Passport)",
                "scheme": "bearer"
            }
        }
    },
...
}